refactor: Remove double/float shim overloads and redundant casts (#151)
* remove double overload for MakeSingleAction

* remove all unnecessary .0

* remove .0f

* remove MakeScale downcast

* remove MakeColor downcast

* remove MakeRotation downcast

* use direct float casting

* use direct float casting pt. 2

* use direct float casting pt. 3

* remove MakePos downcast

* remove AddAnimationDelayLoopFirstLast downcast

* remove unnecessary cast

* remove more cast shims

* remove unnecessary casting

* remove Ceil downcast

* remove Round downcast

* remove downcast for Cosf, Sinf Tanf, Acosf

* use float directly

* remove redundant float castings

* add override keyword

* remove unnecessary double cast

* use MathF

* use mathF pt. 2

* use MathF pt. 3

+++ CutTheRope/Desktop/MouseCursor.cs
@@ -125,8 +125,8 @@ namespace CutTheRope.Desktop
                 for (int x = 0; x < newWidth; x++)
                 {
                     // Map to source coordinates
-                    float srcX = (float)x / newWidth * source.Width;
-                    float srcY = (float)y / newHeight * source.Height;
+                    float srcX = x / newWidth * source.Width;
+                    float srcY = y / newHeight * source.Height;
 
                     // Bilinear interpolation
                     int x0 = Math.Min((int)srcX, source.Width - 1);
