+++ CutTheRope/GameMain/LoadObjects/LoadBouncers.cs
@@ -18,7 +18,7 @@ namespace CutTheRope.GameMain
             float px2 = (xmlNode.AttributeAsNSString("x").IntValue() * scale) + offsetX + mapOffsetX;
             float py2 = (xmlNode.AttributeAsNSString("y").IntValue() * scale) + offsetY + mapOffsetY;
             int w2 = xmlNode.AttributeAsNSString("size").IntValue();
-            double an2 = xmlNode.AttributeAsNSString("angle").IntValue();
+            float an2 = xmlNode.AttributeAsNSString("angle").IntValue();
             Bouncer bouncer = new Bouncer().InitWithPosXYWidthAndAngle(px2, py2, w2, an2);
             bouncer.ParseMover(xmlNode);
             _ = bouncers.AddObject(bouncer);
