+++ CutTheRope/Helpers/TimerManager.cs
@@ -34,11 +34,11 @@ namespace CutTheRope.Helpers
             _ = timers.Remove(timerId);
         }
 
-        public static void RegisterDelayedObjectCall(DelayedDispatcher.DispatchFunc callback, FrameworkTypes parameter, double interval)
+        public static void RegisterDelayedObjectCall(DelayedDispatcher.DispatchFunc callback, FrameworkTypes parameter, float interval)
         {
             ArgumentNullException.ThrowIfNull(callback);
 
-            ArgumentOutOfRangeException.ThrowIfLessThan(interval, 0.0);
+            ArgumentOutOfRangeException.ThrowIfLessThan(interval, 0);
 
             EnsureInitialized();
             delayedDispatcher.CallObjectSelectorParamafterDelay(callback, parameter, interval);
